split a feedable file so I can feed the ai later
