+++ reactive-UI/src/app/canvas-wrapper/managers/UserDrawingLayer.manager.ts
@@ -1,21 +1,10 @@
 import Konva from "konva";
-import { PencilCommands } from "../commands/pencil.command";
-import { ViewPortEventsManager } from "./ViewPortEvents.manager";
 import { Injectable, OnDestroy } from "@angular/core";
-import { StickyNoteCommands } from "../commands/sticky-notes.command";
 import { CursorManager } from "./Cursor.manager";
 import { Subject, debounceTime } from "rxjs";
 import { Group } from "konva/lib/Group";
 import guid from "guid";
 import { Line, LineConfig } from "konva/lib/shapes/Line";
-import { TextInputCommands } from "../commands/text-input.command";
-import { ToBaseEvent, BoardedCreatedEvent, ToDrawingEvent, BaseEvent, StickyNoteMovedEvent, GeneralUndoEvent, AbstractEventQueueItem, StickyNotePastedEvent, InkAttachedToStickyNoteEvent, TextAttachedToStickyNoteEvent, PencilUpEvent, TextEnteredEvent } from "../../events/drawings/EventQueue";
-import { EventsCompositionService } from "../../events/drawings/events-composition.service";
-import { KonvaObjectService } from "../services/3rds/konva-object.service";
-import { KeysService } from "../../services/browser/keys.service";
-import { MetaService } from "../../services/browser/meta.service";
-import { UrlExtractorService } from "../../services/browser/url-extractor.service";
-import { ToolCompositionService } from "../../services/states/tool-composition.service";
 import { SyncingService } from '@com/syncing.service';
 import { PRIMARY_COLOR } from '@config/theme.constants';
 import { FormModalService } from '@ui/controls/form-modal.service';
@@ -23,6 +12,19 @@ import { Point } from '@ui/types/Point';
 import { DrawingObjectService } from '@uidata/drawing-object.service';
 import { EventsService } from '@uidata/events.service';
 import { BoardsService } from '@uidata/boards.service';
+import { KeysService } from '@browser/keys.service';
+import { MetaService } from '@browser/meta.service';
+import { UrlExtractorService } from '@browser/url-extractor.service';
+import { PencilCommands } from '@canvas-module/commands/pencil.command';
+import { StickyNoteCommands } from '@canvas-module/commands/sticky-notes.command';
+import { TextInputCommands } from '@canvas-module/commands/text-input.command';
+import { KonvaObjectService } from '@canvas-module/services/3rds/konva-object.service';
+import { BaseEvent } from '@drawings/BaseEvent';
+import { ToBaseEvent, BoardedCreatedEvent, ToDrawingEvent, StickyNoteMovedEvent, GeneralUndoEvent, StickyNotePastedEvent, InkAttachedToStickyNoteEvent, TextAttachedToStickyNoteEvent, PencilUpEvent, TextEnteredEvent } from '@drawings/EventQueue';
+import { EventsCompositionService } from '@drawings/events-composition.service';
+import { AbstractEventQueueItem } from '@drawings/PureQueue.type';
+import { ToolCompositionService } from '@states/tool-composition.service';
+import { ViewPortEventsManager } from './ViewPortEvents.manager';
 
 @Injectable()
 export class UserDrawingLayerManager implements OnDestroy {
