+++ reactive-UI/src/app/events/drawings/PureQueue.type.ts
@@ -1,0 +1,9 @@
+import { EventCode } from './EventCode';
+
+export type PureQueue = Array<AbstractEventQueueItem>;
+
+export interface AbstractEventQueueItem {
+    code: EventCode;
+    targetId: string;
+    id: string;
+}
