+++ reactive-UI/src/dependencies/socket-communication/web-socket-model.ts
@@ -1,5 +1,5 @@
+import { BaseEvent } from '@drawings/BaseEvent';
 import { UserIdentity } from '@uidata/entities/Identity';
-import { BaseEvent } from "../../app/events/drawings/EventQueue";
 
 export enum WSEventType {
     DRAWING_EVENT = "DRAWING_EVENT",
