+++ reactive-UI/tsconfig.json
@@ -23,7 +23,11 @@
       "@ai/*": ["./src/dependencies/open-ai/*"],
       "@ui/*": ["./src/ui-utilities/*"],
       "@config/*": ["./src/configs/*"],
-      "@uidata/*": ["./src/app/services/data-storages/*"]
+      "@uidata/*": ["./src/app/services/data-storages/*"],
+      "@drawings/*": ["./src/app/events/drawings/*"],
+      "@browser/*": ["./src/app/services/browser/*"],
+      "@canvas-module/*": ["./src/app/canvas-wrapper/*"],
+      "@states/*": ["./src/app/services/states/*"]
     },
     "lib": [
       "ES2022",
